+++ src/video/SDL_yuv_c.h
@@ -22,7 +22,7 @@
 #ifndef SDL_yuv_c_h_
 #define SDL_yuv_c_h_
 
-#include "SDL_internal.h"
+#include "../SDL_internal.h"
 
 /* YUV conversion functions */
 
@@ -31,6 +31,8 @@ extern int SDL_ConvertPixels_RGB_to_YUV(int width, int height, Uint32 src_format
 extern int SDL_ConvertPixels_YUV_to_YUV(int width, int height, Uint32 src_format, const void *src, int src_pitch, Uint32 dst_format, void *dst, int dst_pitch);
 
 
-extern int SDL_CalculateYUVSize(Uint32 format, int w, int h, size_t *size, size_t *pitch);
+extern int SDL_CalculateYUVSize(Uint32 format, int w, int h, size_t *size, int *pitch);
 
 #endif /* SDL_yuv_c_h_ */
+
+/* vi: set ts=4 sw=4 expandtab: */
